pc - refactor town pages into a folder
