+++ frontend/src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import HomePage from "main/pages/HomePage";
-import AvilaBeachPage from "main/pages/AvilaBeachPage";
-import LosAlamosPage from "main/pages/LosAlamosPage";
-import ArroyoGrandePage from "main/pages/ArroyoGrandePage";
+import AvilaBeachPage from "main/pages/Towns/AvilaBeachPage";
+import LosAlamosPage from "main/pages/Towns/LosAlamosPage";
+import ArroyoGrandePage from "main/pages/Towns/ArroyoGrandePage";
 
 import "bootstrap/dist/css/bootstrap.css";
