+++ frontend/src/main/pages/Towns/LosAlamosPage.js
